pyproject.toml
[project]
name = "wattpredictor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "evidently==0.4.31",
    "joblib==1.4.2",
    "lightgbm==4.5.0",
    "matplotlib==3.9.3",
    "numpy==1.26.4",
    "openmeteo-requests==1.3.0",
    "pandas==2.1.4",
    "plotly==5.24.1",
    "pydantic>=1.10.13",
    "pydeck==0.9.1",
    "python-box==7.3.0",
    "python-dotenv==1.0.1",
    "pytz==2024.2",
    "pyyaml==6.0.2",
    "requests==2.32.3",
    "requests-cache==1.2.1",
    "retry-requests==2.0.0",
    "scikit-learn==1.5.2",
    "streamlit==1.40.2",
    "tqdm==4.67.1",
    "xgboost==2.1.3",
]
